backend/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
python-decouple
Pillow
django-cors-headers
gspread
oauth2client
